wip(plugins): Add threading fix
When consuming the states on the dialog it was somehow locking the
main thread and the UI was stuck. In this commit we fixed the
problem and the dialog loads smoother now. The commit is in wip
because we don't know 100% if this is fixed. We need to do more
manual testing