fix(plugins): Fix ui freezing fix
As we haven't discovered yet why exactly the ui is freezing,
we decided to implement a workaround that consists on downloading
the commit types from the resouces using a project sercice and
stoing the result in a singleton. That way the list of commit
types is stored before the dialog is loaded and we can load them
without calling a io operation.